+++ models/tecnologia.js
@@ -1,8 +1,8 @@
 module.exports = function(sequelize, Sequalize) {
     var TechSchema = sequelize.define("Tecnologia", {
-        escritorio: Sequalize.INTEGER,
-        movil: Sequalize.INTEGER,
-        tablet: Sequalize.INTEGER
+        escritorio: Sequalize.FLOAT,
+        movil: Sequalize.FLOAT,
+        tablet: Sequalize.FLOAT
       },{
         freezeTableName: true
     });
